+++ contracts/PrettierParentheses/SubPrettierParentheses.sol
@@ -1,48 +1,91 @@
 pragma solidity 0.5.8;
 
-
 contract SubPrettierParentheses {
-    function subAdd(uint a, uint b, uint c) public pure returns (uint) {
+    function subAdd(uint256 a, uint256 b, uint256 c)
+        public
+        pure
+        returns (uint256)
+    {
         return a - b + c;
     }
 
-    function subSub(uint a, uint b, uint c) public pure returns (uint) {
+    function subSub(uint256 a, uint256 b, uint256 c)
+        public
+        pure
+        returns (uint256)
+    {
         return a - b - c;
     }
 
-    function subMul(uint a, uint b, uint c) public pure returns (uint) {
+    function subMul(uint256 a, uint256 b, uint256 c)
+        public
+        pure
+        returns (uint256)
+    {
         return a - b * c;
     }
 
-    function subDiv(uint a, uint b, uint c) public pure returns (uint) {
+    function subDiv(uint256 a, uint256 b, uint256 c)
+        public
+        pure
+        returns (uint256)
+    {
         return a - b / c;
     }
 
-    function subMod(uint a, uint b, uint c) public pure returns (uint) {
+    function subMod(uint256 a, uint256 b, uint256 c)
+        public
+        pure
+        returns (uint256)
+    {
         return a - (b % c);
     }
 
-    function subExp(uint a, uint b, uint c) public pure returns (uint) {
+    function subExp(uint256 a, uint256 b, uint256 c)
+        public
+        pure
+        returns (uint256)
+    {
         return a - b ** c;
     }
 
-    function subShiftL(uint a, uint b, uint c) public pure returns (uint) {
+    function subShiftL(uint256 a, uint256 b, uint256 c)
+        public
+        pure
+        returns (uint256)
+    {
         return (a - b) << c;
     }
 
-    function subShiftR(uint a, uint b, uint c) public pure returns (uint) {
+    function subShiftR(uint256 a, uint256 b, uint256 c)
+        public
+        pure
+        returns (uint256)
+    {
         return (a - b) >> c;
     }
 
-    function subBitAnd(uint a, uint b, uint c) public pure returns (uint) {
+    function subBitAnd(uint256 a, uint256 b, uint256 c)
+        public
+        pure
+        returns (uint256)
+    {
         return (a - b) & c;
     }
 
-    function subBitOr(uint a, uint b, uint c) public pure returns (uint) {
+    function subBitOr(uint256 a, uint256 b, uint256 c)
+        public
+        pure
+        returns (uint256)
+    {
         return (a - b) | c;
     }
 
-    function subBitXor(uint a, uint b, uint c) public pure returns (uint) {
+    function subBitXor(uint256 a, uint256 b, uint256 c)
+        public
+        pure
+        returns (uint256)
+    {
         return (a - b) ^ c;
     }
 }
